plugins/hex/pyproject.toml
[project]
name = "gestalt-hex"
version = "0.0.1a1"
requires-python = "==3.14.*"
dependencies = ["gestalt-sdk"]

[dependency-groups]
dev = [
  "ruff==0.15.9",
  "ty==0.0.29",
  "vulture==2.16",
]

[tool.uv]
package = false

[tool.uv.sources]
gestalt-sdk = { git = "https://github.com/valon-technologies/gestalt.git", branch = "main", subdirectory = "sdk/python" }

[tool.gestalt]
plugin = "provider"

[tool.gestalt.ci]
vulture-ignore-names = [
  "project_export",
  "project_import",
  "project_run_draft",
  "cell_run",
]
